1. sessionStorage getter and setter were cluttering the useTasksStateManager, so I extracted them to a util file 2. getSessionStorage is now a generic, the return type matches the input type. 3. setSessionStorage is also now generic and return void 4. the tasks and page state now are being set by the getSessionStorage function, much nicer implementation
