+++ src/hooks/use-task-state-manager.ts
@@ -1,5 +1,6 @@
 import { Task, TaskStateManagerProps } from "@/types/task";
 import { useCallback, useEffect, useState } from "react";
+import { getSessionStorage, setSessionStorage } from "./utils/session-storage-utils";
 const TASKS_KEY = 'TASKS';
 const PAGE_KEY = 'PAGE'
 
@@ -7,55 +8,8 @@ const useTaskStateManager = (initialTasks: Task[], initialPage: number): TaskSta
     // multiple stateful values are still being used instead of single object
     // useReducer "makes sense as soon as multiple stateful values are dependent on each other or related to one domain"
 
-    const isBrowser = typeof window != 'undefined';
-    const getSessionStorage = (key: string) => {
-        if (!isBrowser) {
-            if (key === PAGE_KEY) return initialPage;
-            if (key === TASKS_KEY) return initialTasks;
-            throw new Error('Invalid sessionStorage key provided');
-        }
-
-        const val = sessionStorage.getItem(key);
-        if (val) {
-            if (key === PAGE_KEY) return Number(val);
-            if (key === TASKS_KEY) return JSON.parse(val);
-        }
-
-        // no value found in the cache
-        if (key === PAGE_KEY) return initialPage;
-        if (key === TASKS_KEY) return initialTasks;
-        throw new Error('Invalid sessionStorage key provided');
-    }
-    const setSessionStorage = (key: string, val: number | Task[]) => {
-        try {
-            const input = typeof val == 'number' ? val.toString() : JSON.stringify(val);
-            if (isBrowser) sessionStorage.setItem(key, input)
-        } catch (e) {
-            throw new Error(`Failed to store value in sessionStorage: ${e}`);
-        }
-    };
-
-    const [tasks, setTasks] = useState<Task[]>(() => {
-        try {
-            return getSessionStorage(TASKS_KEY);
-        } catch (e) {
-            console.error(`Error retrieving cached tasks: ${e}`);
-            return initialTasks;
-        }
-    });
-
-    const [page, setPage] = useState<number>(() => {
-        try {
-            if (isBrowser) {
-                const cachedPage = sessionStorage.getItem(PAGE_KEY);
-                return cachedPage ? Number(cachedPage) : initialPage;
-            }
-            return initialPage;
-        } catch (e) {
-            console.error(`Error retriveing cached page: ${e}`);
-            return initialPage;
-        }
-    });
+    const [tasks, setTasks] = useState<Task[]>(() => getSessionStorage(TASKS_KEY, initialTasks));
+    const [page, setPage] = useState<number>(() => getSessionStorage(PAGE_KEY, initialPage));
 
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState<string | null>(null);
